receive header folding
Dearest Reviewer,

I looked at issue #81 and found that the CRLF was adding whitespace
to the text during the integer parsing. I added a trim call to the
error flow and try parsing again. I did this because it appears to be
the exceptional case and normal flows will not take the cost of trim
with every process. I can move the trim back out.

I used the failing test that was provided.

Thanks for your time
Becker